requirements_fixed.txt
# 这里面有一大堆错误、冲突、不必要的包，请修复它。
pandas==1.5.3
numpy==1.23.5
requests==2.31.0
openai==1.1.0
tqdm
huggingface_hub==0.17.3
python-dotenv==1.0.0
tenacity==8.2.3  # 用于重试机制